Code Structure Cleaning
