+++ src/Components/Header/Navbar.js
@@ -1,8 +1,7 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { fetchCategories } from '../../Actions/ActionCreators/Categories';
-import { Link } from 'react-router-dom';
-import { skipWelcome } from '../../Actions/ActionCreators/Interface';
+import NavbarCategory from './NavbarCategory';
 
 const Navbar = () => {
   const dispatch = useDispatch();
@@ -16,14 +15,7 @@ const Navbar = () => {
     <nav className='header__navbar navbar'>
       <ul className='navbar__categories'>
         {categories.map((item) => (
-          <li key={item} className='navbar__category'>
-            <Link
-              onClick={() => dispatch(skipWelcome())}
-              to={`/products?category=${item.id}`}
-            >
-              {item.name}
-            </Link>
-          </li>
+          <NavbarCategory item={item} />
         ))}
       </ul>
     </nav>
