+++ src/Components/Header/Searchbar.js
@@ -6,8 +6,7 @@ import {
   fetchSearchProducts,
   setSearchFilter,
 } from '../../Actions/ActionCreators/Products';
-import { Link } from 'react-router-dom';
-import { skipWelcome } from '../../Actions/ActionCreators/Interface';
+import SearchbarResult from './SearchbarResult';
 
 const Searchbar = () => {
   const searchRef = useRef();
@@ -60,32 +59,10 @@ const Searchbar = () => {
           ) : (
             <>
               {results.map((product) => (
-                <div className='search-results__result search-result'>
-                  <img
-                    className='search-result__image'
-                    src={product.image}
-                    alt={product.name}
-                  />
-                  <div className='search-result__info'>
-                    <Link
-                      onClick={() => {
-                        dispatch(skipWelcome());
-                        handleClearProducts();
-                      }}
-                      to={`/products/${product.id}`}
-                    >
-                      <span className='text-weight-400 search-result__name'>
-                        {product.name}
-                      </span>
-                    </Link>
-                    <span className='text-weight-300 search-result__category'>
-                      {product.category.name}
-                    </span>
-                    <span className='text-weight-bold search-result__price'>
-                      {`${product.grossPrice.toFixed(2)}zł`}
-                    </span>
-                  </div>
-                </div>
+                <SearchbarResult
+                  product={product}
+                  handleClearProducts={handleClearProducts}
+                />
               ))}
             </>
           )}
