--- src/Pages/Access/SignUp.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const SignUp = () => {
-  return <div></div>;
-};
-
-export default SignUp;
