--- src/Pages/Cart/DeliveryData.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const DeliveryData = () => {
-  return <div></div>;
-};
-
-export default DeliveryData;
