+++ src/Style/scss/_products-page.scss
@@ -330,6 +330,15 @@
       padding: 5px;
     }
 
+    #{$this}__minimal-amount {
+      font-size: 1.2rem;
+      margin-left: 10px;
+      font-weight: 600;
+      background-color: rgba(0, 0, 0, 0.05);
+      color: rgba(0, 0, 0, 0.4);
+      padding: 5px;
+    }
+
     #{$this}__the-new {
       background-color: rgba(0, 0, 0, 0.1);
       color: $primary-color;
